Rename model_patcher.py -> layer_patcher.py.
