+++ invokeai/backend/patches/layer_patcher.py
